Fixed dyn4j physics engine

+++ engine/src/main/java/tk/timtim3001/engine/physics/colliders/Collider.java
@@ -5,7 +5,7 @@ import org.dyn4j.dynamics.BodyFixture;
 public interface Collider {
 
     BodyFixture getBodyFixture();
-    int getWidth();
-    int getHeight();
+    float getWidth();
+    float getHeight();
 
 }
